gcf-test/Cloud_run_functions/HTTPTriggers/gcf-test/Cloud_run_functions/Event_based_triggers/gcs-to-webpage/requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-cloud-storage
